DTMF based Robotic Car
In this project, the robotic toy car can be controlled using Dual Tone Multi Frequency (DTMF)
technology. This project deals with the operation of a simple toy car that can be controlled by a
cell phone (sender) that makes call to the cell phone (receiver) stacked to the toy car.